continue refactoring: added class CTProcessor. CT class now only contains code for sniffing data files based on filename patterns
